+++ pkg/controllers/dynakube/extension/eec/reconciler.go
@@ -167,3 +167,15 @@ func (r *reconciler) reconcileTlsSecretExpiration(ctx context.Context, secret *c
 	// WIP
 	return nil
 }
+
+func getCertificateAltNames(dkName string) []string {
+	return []string{
+		dkName + "-extensions-controller.dynatrace",
+		dkName + "-extensions-controller.dynatrace.svc",
+		dkName + "-extensions-controller.dynatrace.svc.cluster.local",
+	}
+}
+
+func getTlsSecretName(dkName string) string {
+	return dkName + consts.ExtensionsTlsSecretSuffix
+}
